refactor code to add modules/ directory
